Wed 18 Aug, Lesson 3
Bug Fixes & Changes
The problem of the ball not rolling in Level 2 has been fixed, turns out there was some error with the physics and the GameComponent. Replacing the ball from Level 2 to Level 1 fixes it.
Added new obstacles and a 'track' to Level 2
Fixed the checkpoint in Level 1, so that every time the ball passes through it updates the myXYZ variables, so that it will automatically save.
For Level 2, I have increased the wind so that the grass and trees look more realistic
Added sound effects and background nature sounds